requirements.txt
Flask==3.1.2
Werkzeug==3.1.3
pymongo==4.6.0
python-dotenv==1.0.0
requests==2.32.5

